# zstd: Optimize copyMemoryPrecise further