The 16-byte block copy loop is one instruction shorter and all the pointer updating is done at the end to decrease the code size.

This gives a small speedup on some benchmarks, with smaller regressions on some of the others:

```
name                                                   old speed      new speed      delta
Decoder_DecoderSmall/kppkn.gtb.zst/buffered-16          714MB/s ± 1%   719MB/s ± 3%    ~     (p=1.000 n=9+10)
Decoder_DecoderSmall/kppkn.gtb.zst/unbuffered-16       1.16GB/s ± 8%  1.19GB/s ± 4%    ~     (p=0.247 n=10+10)
Decoder_DecoderSmall/geo.protodata.zst/buffered-16     2.25GB/s ± 3%  2.33GB/s ± 4%  +3.30%  (p=0.007 n=9+10)
Decoder_DecoderSmall/geo.protodata.zst/unbuffered-16   2.39GB/s ± 0%  2.51GB/s ± 2%  +5.28%  (p=0.000 n=10+10)
Decoder_DecoderSmall/plrabn12.txt.zst/buffered-16       512MB/s ± 2%   519MB/s ± 3%  +1.45%  (p=0.019 n=10+10)
Decoder_DecoderSmall/plrabn12.txt.zst/unbuffered-16    1.18GB/s ± 5%  1.16GB/s ± 6%    ~     (p=0.393 n=10+10)
Decoder_DecoderSmall/lcet10.txt.zst/buffered-16         612MB/s ± 2%   614MB/s ± 3%    ~     (p=0.796 n=10+10)
Decoder_DecoderSmall/lcet10.txt.zst/unbuffered-16      1.21GB/s ± 8%  1.22GB/s ± 4%    ~     (p=0.912 n=10+10)
Decoder_DecoderSmall/asyoulik.txt.zst/buffered-16       553MB/s ± 2%   553MB/s ± 3%    ~     (p=0.631 n=10+10)
Decoder_DecoderSmall/asyoulik.txt.zst/unbuffered-16    1.16GB/s ± 5%  1.19GB/s ± 4%  +3.12%  (p=0.035 n=10+10)
Decoder_DecoderSmall/alice29.txt.zst/buffered-16        553MB/s ± 2%   555MB/s ± 1%    ~     (p=0.182 n=10+9)
Decoder_DecoderSmall/alice29.txt.zst/unbuffered-16      941MB/s ± 2%   927MB/s ± 4%    ~     (p=0.143 n=10+10)
Decoder_DecoderSmall/html_x_4.zst/buffered-16          4.06GB/s ± 2%  4.10GB/s ± 3%    ~     (p=0.105 n=10+10)
Decoder_DecoderSmall/html_x_4.zst/unbuffered-16        4.71GB/s ± 1%  4.72GB/s ± 2%    ~     (p=0.247 n=10+10)
Decoder_DecoderSmall/paper-100k.pdf.zst/buffered-16    5.96GB/s ± 4%  5.77GB/s ± 4%  -3.21%  (p=0.006 n=10+9)
Decoder_DecoderSmall/paper-100k.pdf.zst/unbuffered-16  5.70GB/s ± 1%  5.80GB/s ± 1%  +1.69%  (p=0.000 n=10+10)
Decoder_DecoderSmall/fireworks.jpeg.zst/buffered-16    11.4GB/s ± 3%  11.6GB/s ± 2%    ~     (p=0.123 n=10+10)
Decoder_DecoderSmall/fireworks.jpeg.zst/unbuffered-16  9.76GB/s ± 1%  9.81GB/s ± 2%    ~     (p=0.400 n=9+10)
Decoder_DecoderSmall/urls.10K.zst/buffered-16           829MB/s ± 2%   815MB/s ± 2%  -1.68%  (p=0.007 n=10+10)
Decoder_DecoderSmall/urls.10K.zst/unbuffered-16        1.48GB/s ± 1%  1.45GB/s ± 3%  -1.88%  (p=0.007 n=10+10)
Decoder_DecoderSmall/html.zst/buffered-16              1.72GB/s ± 1%  1.77GB/s ± 2%  +3.09%  (p=0.000 n=9+10)
Decoder_DecoderSmall/html.zst/unbuffered-16            2.10GB/s ± 2%  2.16GB/s ± 1%  +2.67%  (p=0.000 n=9+10)
Decoder_DecoderSmall/comp-data.bin.zst/buffered-16      684MB/s ± 2%   683MB/s ± 2%    ~     (p=0.905 n=9+10)
Decoder_DecoderSmall/comp-data.bin.zst/unbuffered-16    617MB/s ± 1%   614MB/s ± 1%    ~     (p=0.123 n=10+10)
```