@greatroar 2022-11-22
I don't understand the error. When I run go test (w/ Go 1.19), everything passes. But when I run

    go test -run=TestEncoderRegression/fastest-c1-w1k-1seg

I also get that "panic: test executed panic(nil) or runtime.Goexit".