@greatroar 2022-11-22
Still failing, even though the tests pass locally now. I'm also seeing performance regressions up to 7% on an older i7.